+++ OptimizelySDK.Tests/EventTests/EventBuilderTest.cs
@@ -32,6 +32,7 @@ namespace OptimizelySDK.Tests.EventTests
     {
         private string TestUserId = string.Empty;
         private ProjectConfig Config;
+
         private EventBuilder EventBuilder;
 
         [TestFixtureSetUp]
@@ -83,7 +84,7 @@ namespace OptimizelySDK.Tests.EventTests
                                     }
                                 }
                             },
-                            {"attributes", new object[] 
+                            {"attributes", new object[]
                                 {
                                     new Dictionary<string, object>
                                     {
@@ -107,7 +108,7 @@ namespace OptimizelySDK.Tests.EventTests
             };
 
             var expectedLogEvent = new LogEvent("https://logx.optimizely.com/v1/events",
-                payloadParams, 
+                payloadParams,
                 "POST",
                 new Dictionary<string, string>
                 {
@@ -191,10 +192,10 @@ namespace OptimizelySDK.Tests.EventTests
                 {"revision", "15" },
                 {"anonymize_ip", false}
             };
-            
+
             var expectedLogEvent = new LogEvent("https://logx.optimizely.com/v1/events",
                 payloadParams,
-                "POST", 
+                "POST",
                 new Dictionary<string, string>
                 {
                     { "Content-Type", "application/json" }
@@ -209,6 +210,228 @@ namespace OptimizelySDK.Tests.EventTests
 
             var logEvent = EventBuilder.CreateImpressionEvent(Config, Config.GetExperimentFromKey("test_experiment"), "77210100090", TestUserId, userAttributes);
 
+            TestData.ChangeGUIDAndTimeStamp(logEvent.Params, timeStamp, guid);
+
+            Assert.IsTrue(TestData.CompareObjects(expectedLogEvent, logEvent));
+        }
+
+        [Test]
+        public void TestCreateImpressionEventWithTypedAttributes()
+        {
+            var guid = Guid.NewGuid();
+            var timeStamp = TestData.SecondsSince1970();
+
+            var payloadParams = new Dictionary<string, object>
+            {
+                { "visitors", new object[]
+                    {
+                        new Dictionary<string, object>()
+                        {
+                            { "snapshots", new object[]
+                                {
+                                    new Dictionary<string, object>
+                                    {
+                                        { "decisions", new object[]
+                                            {
+                                                new Dictionary<string, object>
+                                                {
+                                                    {"campaign_id", "7719770039" },
+                                                    {"experiment_id", "7716830082" },
+                                                    {"variation_id", "7722370027" }
+                                                }
+                                            }
+                                        },
+                                        { "events", new object[]
+                                            {
+                                                new Dictionary<string, object>
+                                                {
+                                                    {"entity_id", "7719770039" },
+                                                    {"timestamp", timeStamp },
+                                                    {"uuid", guid },
+                                                    {"key", "campaign_activated" }
+                                                }
+                                            }
+                                        }
+                                    }
+                                }
+                            },
+                            {"attributes", new object[]
+                                {
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", "7723280020" },
+                                        {"key", "device_type" },
+                                        {"type", "custom" },
+                                        {"value", "iPhone"}
+                                    },
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", "323434545" },
+                                        {"key", "boolean_key" },
+                                        {"type", "custom" },
+                                        {"value", true}
+                                    },
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", "616727838" },
+                                        {"key", "integer_key" },
+                                        {"type", "custom" },
+                                        {"value", 15}
+                                    },
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", "808797686" },
+                                        {"key", "double_key" },
+                                        {"type", "custom" },
+                                        {"value", 3.14}
+                                    },
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", ControlAttributes.BOT_FILTERING_ATTRIBUTE},
+                                        {"key", ControlAttributes.BOT_FILTERING_ATTRIBUTE},
+                                        {"type", "custom" },
+                                        {"value", true }
+                                    }
+                                }
+                            },
+                            { "visitor_id", TestUserId }
+                        }
+                    }
+                },
+                {"project_id", "7720880029" },
+                {"account_id", "1592310167" },
+                {"client_name", "csharp-sdk" },
+                {"client_version", Optimizely.SDK_VERSION },
+                {"revision", "15" },
+                {"anonymize_ip", false}
+            };
+
+            var expectedLogEvent = new LogEvent("https://logx.optimizely.com/v1/events",
+                payloadParams,
+                "POST",
+                new Dictionary<string, string>
+                {
+                    { "Content-Type", "application/json" }
+                });
+
+            var userAttributes = new UserAttributes
+            {
+                {"device_type", "iPhone" },
+                {"boolean_key", true },
+                {"integer_key", 15 },
+                {"double_key", 3.14 }
+            };
+
+            var logEvent = EventBuilder.CreateImpressionEvent(Config, Config.GetExperimentFromKey("test_experiment"), "7722370027", TestUserId, userAttributes);
+            TestData.ChangeGUIDAndTimeStamp(logEvent.Params, timeStamp, guid);
+
+            Assert.IsTrue(TestData.CompareObjects(expectedLogEvent, logEvent));
+        }
+
+        [Test]
+        public void TestCreateImpressionEventRemovesInvalidAttributesFromPayload()
+        {
+            var guid = Guid.NewGuid();
+            var timeStamp = TestData.SecondsSince1970();
+
+            var payloadParams = new Dictionary<string, object>
+            {
+                { "visitors", new object[]
+                    {
+                        new Dictionary<string, object>()
+                        {
+                            { "snapshots", new object[]
+                                {
+                                    new Dictionary<string, object>
+                                    {
+                                        { "decisions", new object[]
+                                            {
+                                                new Dictionary<string, object>
+                                                {
+                                                    {"campaign_id", "7719770039" },
+                                                    {"experiment_id", "7716830082" },
+                                                    {"variation_id", "7722370027" }
+                                                }
+                                            }
+                                        },
+                                        { "events", new object[]
+                                            {
+                                                new Dictionary<string, object>
+                                                {
+                                                    {"entity_id", "7719770039" },
+                                                    {"timestamp", timeStamp },
+                                                    {"uuid", guid },
+                                                    {"key", "campaign_activated" }
+                                                }
+                                            }
+                                        }
+                                    }
+                                }
+                            },
+                            {"attributes", new object[]
+                                {
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", "7723280020" },
+                                        {"key", "device_type" },
+                                        {"type", "custom" },
+                                        {"value", "iPhone"}
+                                    },
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", "323434545" },
+                                        {"key", "boolean_key" },
+                                        {"type", "custom" },
+                                        {"value", true}
+                                    },
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", "808797686" },
+                                        {"key", "double_key" },
+                                        {"type", "custom" },
+                                        {"value", 3.14}
+                                    },
+                                    new Dictionary<string, object>
+                                    {
+                                        {"entity_id", ControlAttributes.BOT_FILTERING_ATTRIBUTE},
+                                        {"key", ControlAttributes.BOT_FILTERING_ATTRIBUTE},
+                                        {"type", "custom" },
+                                        {"value", true }
+                                    }
+                                }
+                            },
+                            { "visitor_id", TestUserId }
+                        }
+                    }
+                },
+                {"project_id", "7720880029" },
+                {"account_id", "1592310167" },
+                {"client_name", "csharp-sdk" },
+                {"client_version", Optimizely.SDK_VERSION },
+                {"revision", "15" },
+                {"anonymize_ip", false}
+            };
+
+            var expectedLogEvent = new LogEvent("https://logx.optimizely.com/v1/events",
+                payloadParams,
+                "POST",
+                new Dictionary<string, string>
+                {
+                    { "Content-Type", "application/json" }
+                });
+
+            var userAttributes = new UserAttributes
+            {
+                {"device_type", "iPhone" },
+                {"boolean_key", true },
+                {"double_key", 3.14 },
+                { "", "Android" },
+                { "null", null },
+                { "objects", new object() },
+                { "arrays", new string[] { "a", "b", "c" } },
+            };
+
+            var logEvent = EventBuilder.CreateImpressionEvent(Config, Config.GetExperimentFromKey("test_experiment"), "7722370027", TestUserId, userAttributes);
             TestData.ChangeGUIDAndTimeStamp(logEvent.Params, timeStamp, guid);
 
             Assert.IsTrue(TestData.CompareObjects(expectedLogEvent, logEvent));
@@ -255,7 +478,7 @@ namespace OptimizelySDK.Tests.EventTests
                                 }
                             },
                             {"visitor_id", TestUserId },
-                            {"attributes", new object[] 
+                            {"attributes", new object[]
                                 {
                                     new Dictionary<string, object>
                                     {
@@ -276,7 +499,7 @@ namespace OptimizelySDK.Tests.EventTests
                 {"revision", "15" },
                 {"anonymize_ip", false}
             };
-            
+
             var expectedEvent = new LogEvent(
                 "https://logx.optimizely.com/v1/events",
                 payloadParams,
@@ -302,7 +525,7 @@ namespace OptimizelySDK.Tests.EventTests
         {
             var guid = Guid.NewGuid();
             var timeStamp = TestData.SecondsSince1970();
-            
+
             var payloadParams = new Dictionary<string, object>
             {
                 {"visitors", new object[]
@@ -597,7 +820,7 @@ namespace OptimizelySDK.Tests.EventTests
 
             Assert.IsTrue(TestData.CompareObjects(expectedEvent, logEvent));
         }
-        
+
         [Test]
         public void TestCreateConversionEventNoAttributesWithInvalidValue()
         {
@@ -668,7 +891,7 @@ namespace OptimizelySDK.Tests.EventTests
                 {"revision", "15" },
                 {"anonymize_ip", false}
             };
-            
+
             var expectedEvent = new LogEvent(
                 "https://logx.optimizely.com/v1/events",
                 payloadParams,
@@ -989,7 +1212,7 @@ namespace OptimizelySDK.Tests.EventTests
                 { "company", "Optimizely" },
                 {ControlAttributes.BUCKETING_ID_ATTRIBUTE, "variation" }
             };
-            
+
             var logEvent = EventBuilder.CreateImpressionEvent(Config, Config.GetExperimentFromKey("test_experiment"), "7722370027", TestUserId, userAttributes);
 
             TestData.ChangeGUIDAndTimeStamp(logEvent.Params, timeStamp, guid);
@@ -1361,6 +1584,140 @@ namespace OptimizelySDK.Tests.EventTests
             TestData.ChangeGUIDAndTimeStamp(logEvent.Params, timeStamp, guid);
 
             Assert.IsTrue(TestData.CompareObjects(expectedEvent, logEvent));
+        }
+
+        [Test]
+        public void TestCreateConversionEventWhenEventUsedInMultipleExp()
+        {
+            var guid = Guid.NewGuid();
+            var timeStamp = TestData.SecondsSince1970();
+            
+            var eventInMultiExperimentConfig = ProjectConfig.Create(TestData.SimpleABExperimentsDatafile, new NoOpLogger(), new ErrorHandler.NoOpErrorHandler());
+
+            var experimentIdVariationMap = new Dictionary<string, Variation>
+            {
+                {
+                    "111127", new Variation{Id="111129", Key="variation"}
+                },
+                {
+                    "111130", new Variation{Id="111131", Key="variation"}
+                }
+            };
+
+            var logEvent = EventBuilder.CreateConversionEvent(eventInMultiExperimentConfig, "event_with_multiple_running_experiments", experimentIdVariationMap, "test_user",
+                                                              new UserAttributes {
+                                                                {"test_attribute", "test_value"}
+                                                              },
+                                                              new EventTags {
+                                                                {"revenue", 4200},
+                                                                {"value", 1.234},
+                                                                {"non-revenue", "abc"}
+                                                             });
+                    
+            var payloadParams = new Dictionary<string, object>
+                {
+                {"client_version", Optimizely.SDK_VERSION},
+                {"project_id", "111001"},
+                //{"visitor_id", "test_user"},
+                {"account_id", "12001"},
+                {"client_name", "csharp-sdk"},
+                {"anonymize_ip", false},
+                {"revision", eventInMultiExperimentConfig.Revision},
+                {"visitors", new object[]
+                    {
+                        //visitors[0]
+                        new Dictionary<string, object>
+                        {
+                            //visitors[0].attributes
+                            {
+                                "attributes", new object[]
+                                {
+                                    new Dictionary<string, string>
+                                    {
+                                        {"entity_id", "111094"},
+                                        {"type", "custom"},
+                                        {"value", "test_value"},
+                                        {"key", "test_attribute"}
+                                    }
+                                }
+                            },
+                            //visitors[0].visitor_id
+                            {"visitor_id", "test_user"},
+                            //visitors[0].snapshots
+                            {"snapshots", new object[]
+                                {
+                                    //snapshots[0]
+                                    new Dictionary<string, object>
+                                    {
+                                        //snapshots[0].decisions
+                                        {"decisions", new object[]
+                                            {
+                                                //decisions[0]
+                                                new Dictionary<string, object>
+                                                {
+                                                    {"variation_id", "111129"},
+                                                    {"experiment_id", "111127"},
+                                                    {"campaign_id", "111182"}
+
+                                                },
+                                                //decisions[1]
+                                                new Dictionary<string,object>
+                                                {
+                                                    {"experiment_id", "111130"},
+                                                    {"variation_id", "111131"},
+                                                    {"campaign_id", "111182"}
+                                                }
+                                            }
+                                        },
+                                        //snapshots[0].events
+                                        {
+                                            "events", new object[]
+                                            {
+                                                new Dictionary<string, object>
+                                                {
+                                                    {"uuid", guid},
+                                                    {"timestamp", timeStamp},
+                                                    {"revenue", 4200},
+                                                    {"value", 1.234},
+                                                    {"key", "event_with_multiple_running_experiments"},
+                                                    {"entity_id", "111095"},
+                                                    {
+                                                        "tags", new Dictionary<string, object>
+                                                        {
+                                                            {"non-revenue", "abc"},
+                                                            {"revenue", 4200},
+                                                            {"value", 1.234},
+                                                        }
+                                                    }
+
+                                                }
+                                            }
+                                        }
+
+                                    }
+
+                                }
+                            }
+
+                        }
+                    }
+
+                }
+            };
+
+
+            var expectedLogEvent = new LogEvent(
+                "https://logx.optimizely.com/v1/events",
+                payloadParams,
+                "POST",
+                new Dictionary<string, string>
+                {
+                    { "Content-Type", "application/json"}
+                });
+
+            TestData.ChangeGUIDAndTimeStamp(logEvent.Params, timeStamp, guid);
+
+            Assert.IsTrue(TestData.CompareObjects(expectedLogEvent, logEvent));
         }
     }
 }