+++ OptimizelySDK.Tests/ProjectConfigTest.cs
@@ -117,7 +117,10 @@ namespace OptimizelySDK.Tests
             {
                 { "device_type", Config.GetAttribute("device_type") },
                 { "location", Config.GetAttribute("location")},
-                { "browser_type", Config.GetAttribute("browser_type")}
+                { "browser_type", Config.GetAttribute("browser_type")},
+                { "boolean_key", Config.GetAttribute("boolean_key")},
+                { "integer_key", Config.GetAttribute("integer_key")},
+                { "double_key", Config.GetAttribute("double_key")}
             };
             Assert.IsTrue(TestData.CompareObjects(attributeKeyMap, Config.AttributeKeyMap));
 
@@ -862,6 +865,33 @@ namespace OptimizelySDK.Tests
             // Verify that null is returned when provided attribute key is invalid.
             Assert.Null(Config.GetAttributeId("invalid_attribute"));
             LoggerMock.Verify(l => l.Log(LogLevel.ERROR, @"Attribute key ""invalid_attribute"" is not in datafile."));
+        }
+
+        [Test]
+        public void TestCreateThrowsWithNullDatafile()
+        {
+            var exception = Assert.Throws<ConfigParseException>(() => ProjectConfig.Create(null, null, null));
+            Assert.AreEqual("Unable to parse null datafile.", exception.Message);
+        }
+
+        [Test]
+        public void TestCreateThrowsWithEmptyDatafile()
+        {
+            var exception = Assert.Throws<ConfigParseException>(() => ProjectConfig.Create("", null, null));
+            Assert.AreEqual("Unable to parse empty datafile.", exception.Message);
+        }
+
+        [Test]
+        public void TestCreateThrowsWithUnsupportedDatafileVersion()
+        {
+            var exception = Assert.Throws<ConfigParseException>(() => ProjectConfig.Create(TestData.UnsupportedVersionDatafile, null, null));
+            Assert.AreEqual($"This version of the C# SDK does not support the given datafile version: 5", exception.Message);
+        }
+
+        [Test]
+        public void TestCreateDoesNotThrowWithValidDatafile()
+        {
+            Assert.DoesNotThrow(() => ProjectConfig.Create(TestData.Datafile, null, null));
         }
     }
 }