+++ OptimizelySDK/Entity/UserAttributes.cs
@@ -1,5 +1,5 @@
 ﻿/* 
- * Copyright 2017, Optimizely
+ * Copyright 2017-2018, Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,19 +18,8 @@ using OptimizelySDK.Logger;
 
 namespace OptimizelySDK.Entity
 {
-    public class UserAttributes : Dictionary<string, string>
+    public class UserAttributes : Dictionary<string, object>
     {
-        public UserAttributes FilterNullValues(ILogger logger)
-        {
-            UserAttributes answer = new UserAttributes();
-            foreach (KeyValuePair<string, string> pair in this) {
-                if (pair.Value != null) {
-                    answer[pair.Key] = pair.Value;
-                } else {
-                    logger.Log(LogLevel.ERROR, string.Format("[UserAttributes] Null value for key {0} removed and will not be sent to results.", pair.Key));
-                }
-            }
-            return answer;
-        }
+        
     }
 }